server/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.1
requests==2.30.0
python-dotenv==1.0.0
firebase-admin==6.3.0